requirements.txt
assemblyai>=0.63.0
anthropic>=0.49.0
openai>=1.30.0
python-dotenv==1.0.1
pydub==0.25.1
rich==13.9.4
streamlit>=1.56.0
bcrypt>=4.0.0
pyyaml>=6.0
supabase>=2.3.0
